- Warning fix, I guess
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@2447 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Quests.cpp
@@ -73,7 +73,7 @@ bool Quests::giveMesos(Player *player, int32_t amount) {
 }
 
 void Quests::giveFame(Player *player, int32_t amount) {
-	player->setFame(player->getFame() + amount);
+	player->setFame(player->getFame() + static_cast<int16_t>(amount));
 	QuestsPacket::giveFame(player, amount);
 }
